перенёс интерфейсы в свою папку
